refactor(#9). 구조에 맞게 파일 위치 반영
