+++ src/entities/mission/model/mock-mission.tsx
