+++ src/entities/user/model/mock-user.tsx
