+++ src/entities/user/model/type.ts
