+++ src/features/mission/ui/mission-join-modal.tsx
@@ -1,3 +1,4 @@
+import { JoinMissionRequest } from '@/entities/mission/model/type';
 import { useJoinMission } from '@/features/mission/api/use-mission-service';
 import { Button } from '@/shared/ui/button';
 import {
@@ -7,7 +8,6 @@ import {
   DialogHeader,
   DialogTitle,
 } from '@/shared/ui/dialog';
-import { JoinMissionRequest } from '@/types/mission';
 import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 import { LuLoader2 } from 'react-icons/lu';
